removed the creating of letsencrypt/live dir as when we mount the volume this path is overwritten with the volume. Added logic to check that letsencrypt/live exists before we try to load certificates from a non-existant path. change printerr to print as this was causing logging to be incorrectly ordered making it hard to read the logs.

+++ shared/lib/src/certbot/certbot.dart
@@ -58,61 +58,74 @@ class Certbot {
       @required bool autoAcquireMode}) {
     var hasValidCerts = false;
 
-    envs.forEach((key, value) => print('$key=$value'));
+    /// envs.forEach((key, value) => print('$key=$value'));
 
-    var certs = certificates();
+    /// If there is no live path then no certificates have been acquired.
+    if (exists(CertbotPaths.letsEncryptLivePath)) {
+      var certs = certificates();
 
-    for (var cert in certs) {
-      print(cert.toString());
-    }
+      for (var cert in certs) {
+        print(cert.toString());
+      }
 
-    if (!revoking) {
-      var path = CertbotPaths.fullChainPath(CertbotPaths.certificatePathRoot(hostname, domain, wildcard: wildcard));
-      if (exists(path)) {
-        print('Found fullchain in: $path');
+      if (!revoking) {
+        var path = CertbotPaths.fullChainPath(CertbotPaths.certificatePathRoot(
+            hostname, domain,
+            wildcard: wildcard));
+        if (exists(path)) {
+          print('Found fullchain in: $path');
 
-        if (hasExpired(hostname, domain)) {
-          printerr("ERROR The Certificate for $hostname.$domain has expired. Please run 'nginx-le acquire.");
+          if (hasExpired(hostname, domain)) {
+            print(
+                "ERROR The Certificate for $hostname.$domain has expired. Please run 'nginx-le acquire.");
+          } else {
+            hasValidCerts = true;
+          }
         } else {
-          hasValidCerts = true;
-        }
-      } else {
-        if (!autoAcquireMode) {
-          printerr("No Certificates found for $hostname.$domain. You may need to run 'nginx-le acquire");
+          if (!autoAcquireMode) {
+            print(
+                "No Certificates found for $hostname.$domain. You may need to run 'nginx-le acquire");
+          }
         }
       }
-    }
 
-    if (exists(LIVE_WWW_PATH, followLinks: false)) {
-      deleteSymlink(LIVE_WWW_PATH);
+      if (exists(LIVE_WWW_PATH, followLinks: false)) {
+        deleteSymlink(LIVE_WWW_PATH);
+      }
     }
 
     if (hasValidCerts) {
       print(orange('Deploying certificates'));
 
       /// symlink the user's custom content.
+      print('symlinking /etc/nginx/custom');
       symlink('/etc/nginx/custom', LIVE_WWW_PATH);
-      _deploy(CertbotPaths.certificatePathRoot(hostname, domain, wildcard: wildcard));
-      printerr(green('*') * 120);
-      printerr(green('* Nginx-LE is running with an active Certificate.'));
-      printerr(green('*') * 120);
+      _deploy(CertbotPaths.certificatePathRoot(hostname, domain,
+          wildcard: wildcard));
+      print(green('*') * 120);
+      print(green('* Nginx-LE is running with an active Certificate.'));
+      print(green('*') * 120);
     } else {
-      printerr(red('*') * 120);
-      printerr(red('No certificates Found!!'));
-      printerr(red(
+      print(red('*') * 120);
+      print(red('No certificates Found during deploy!!'));
+      print(red(
           "* Nginx-LE is running in 'Certificate Acquisition' mode. It will only respond to CertBot validation requests."));
-      printerr(red('*') * 120);
+      print(red('*') * 120);
 
       /// symlink in the http configs which only permit certbot access
+      print('symlinking /etc/nginx/acquire');
       symlink('/etc/nginx/acquire', LIVE_WWW_PATH);
     }
 
     if (reload) {
       _reloadNginx();
     }
+    print('Deploy complete.');
+    sleep(5);
   }
 
-  void deployCertificatesDirect(String certificateRootPath, {bool revoking = false}) {
+  void deployCertificatesDirect(String certificateRootPath,
+      {bool revoking = false}) {
     if (exists(LIVE_WWW_PATH, followLinks: false)) {
       deleteSymlink(LIVE_WWW_PATH);
     }
@@ -144,14 +157,18 @@ class Certbot {
   void _deploy(String certificateRootPath) {
     /// we need to leave the original files in place as they form part
     /// of the letsencrypt archive
-    copy(CertbotPaths.fullChainPath(certificateRootPath), '/tmp/fullchain.pem', overwrite: true);
-    copy(CertbotPaths.privateKeyPath(certificateRootPath), '/tmp/privkey.pem', overwrite: true);
+    copy(CertbotPaths.fullChainPath(certificateRootPath), '/tmp/fullchain.pem',
+        overwrite: true);
+    copy(CertbotPaths.privateKeyPath(certificateRootPath), '/tmp/privkey.pem',
+        overwrite: true);
 
     /// but we need to move them in place using move so that
     /// the replace is essentially atomic so that nginx doesn't see partially
     /// created certificates.
-    move('/tmp/fullchain.pem', join(nginxCertPath, 'fullchain.pem'), overwrite: true);
-    move('/tmp/privkey.pem', join(nginxCertPath, 'privkey.pem'), overwrite: true);
+    move('/tmp/fullchain.pem', join(nginxCertPath, 'fullchain.pem'),
+        overwrite: true);
+    move('/tmp/privkey.pem', join(nginxCertPath, 'privkey.pem'),
+        overwrite: true);
   }
 
   /// Used more for testing, but essentially deletes any existing certificates
@@ -180,7 +197,7 @@ class Certbot {
 
     var progress = Progress(
       (line) => print(line),
-      stderr: (line) => printerr(line),
+      stderr: (line) => print(line),
     );
     cmd.start(
       runInShell: true,
@@ -197,7 +214,8 @@ class Certbot {
 
   /// used by revoke to delete certificates after they have been revoked
   /// If we don't do this then the revoked certificates will still be renewed.
-  void _delete(String hostname, String domain, {@required String emailaddress}) {
+  void _delete(String hostname, String domain,
+      {@required String emailaddress}) {
     var workDir = _createDir(CertbotPaths.letsEncryptWorkPath);
     var logDir = _createDir(CertbotPaths.letsEncryptLogPath);
     var configDir = _createDir(CertbotPaths.letsEncryptConfigPath);
@@ -215,10 +233,11 @@ class Certbot {
         runInShell: true,
         nothrow: true,
         progress: Progress((line) {
-          if (!line.startsWith('- - - -') && !line.startsWith('Saving debug ')) {
+          if (!line.startsWith('- - - -') &&
+              !line.startsWith('Saving debug ')) {
             print(line);
           }
-        }, stderr: (line) => printerr(line)));
+        }, stderr: (line) => print(line)));
   }
 
   /// Checks if the certificate for the given hostname.domain
@@ -252,15 +271,16 @@ class Certbot {
       try {
         renew();
       } on CertbotException catch (e, st) {
-        printerr(e.message);
-        printerr(e.details);
-        printerr(st.toString());
-        Email.sendError(subject: e.message, body: '${e.details}\n ${st.toString()}');
+        print(e.message);
+        print(e.details);
+        print(st.toString());
+        Email.sendError(
+            subject: e.message, body: '${e.details}\n ${st.toString()}');
       } catch (e, st) {
         /// we don't rethrow as we don't want to shutdown the scheduler.
         /// as this may be a temporary error.
-        printerr(e.toString());
-        printerr(st.toString());
+        print(e.toString());
+        print(st.toString());
 
         Email.sendError(subject: e.toString(), body: st.toString());
       }
@@ -281,7 +301,7 @@ class Certbot {
       print(line);
       lines.add(line);
     }, stderr: (line) {
-      printerr(line);
+      print(line);
       lines.add(line);
     });
 
@@ -290,7 +310,8 @@ class Certbot {
     if (progress.exitCode != 0) {
       var system = 'hostname'.firstLine;
 
-      throw CertbotException('certbot failed renewing a certificate for ${Environment().fqdn}on $system',
+      throw CertbotException(
+          'certbot failed renewing a certificate for ${Environment().fqdn}on $system',
           details: lines.join('\n'));
     }
   }
@@ -375,7 +396,10 @@ class Certbot {
   bool get _hasValidBlockFlag {
     var valid = false;
     if (exists(_pathToBlockFlag)) {
-      valid = stat(_pathToBlockFlag).changed.add(Duration(minutes: 15)).isBefore(DateTime.now());
+      valid = stat(_pathToBlockFlag)
+          .changed
+          .add(Duration(minutes: 15))
+          .isBefore(DateTime.now());
     }
     return valid;
   }
@@ -386,7 +410,8 @@ class Certbot {
     }
   }
 
-  String get _pathToBlockFlag => join(Environment().certbotRootPath, 'block_acquisitions.flag');
+  String get _pathToBlockFlag =>
+      join(Environment().certbotRootPath, 'block_acquisitions.flag');
 }
 
 class CertbotException implements Exception {
